% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkgsampling.R
\name{bkgsampling}
\alias{bkgsampling}
\title{Background points training and testing dataset sampling}
\usage{
bkgsampling(
  env.rast,
  pres = NULL,
  thres = 0.75,
  H = NULL,
  grid.res = NULL,
  n.tr = 5,
  sub.ts = FALSE,
  n.ts = 5,
  plot_proc = FALSE
)
}
\arguments{
\item{env.rast}{A raster stack or raster brick object comprising the variables describing the environmental space.}

\item{pres}{A SpatialPointDataFrame including species of interest presence-only.}

\item{thres}{Kernel density threshold SPIEGARE}

\item{H}{Parameter defining kernel bandwidth, it can be provided by the user or automatically estimated via SPIEGARE}

\item{grid.res}{(integer) resolution of the sampling grid. The resolution can be arbitrarily selected or defined using the Optim_res() function.}

\item{n.tr}{(integer) number of background points for the training dataset to sample in each cell of the sampling grid}

\item{sub.ts}{(logical) sample the validation background points}

\item{n.ts}{(integer; optional) number of background points for the testing dataset to sample in each cell of the sampling grid. sub.ts argument must be TRUE.}

\item{plot_proc}{(logical) plot progress of the sampling, default FALSE}
}
\value{
A spatial point data frame with the coordinates of the background points both in the geographical and environmental space.
}
\description{
This function internally calls the uesampling() function to perform the uniform sampling of the environmental space.
Before performing the uniform sampling of the environmental space, a set of PC-scores within a core area associated with the species presences
is filtered out using a kernel density estimation approach.
The bandwith of the kernel can be either defined by the user or automatically estimated, allowing the user to account for metapopulation dynamics (e.g. sink-source populations).
}
