% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_res.R
\name{optim_res}
\alias{optim_res}
\title{Get optimal resolution of the sampling grid}
\usage{
optim_res(sdf, grid.res, perc.thr = 10, showOpt = TRUE)
}
\arguments{
\item{sdf}{an sf object having point geometry given by the PC-scores values}

\item{grid.res}{(integer) a vector of resolutions to be tested, i.e seq(1,100, by=1)}

\item{perc.thr}{rate of change (expressed in percentage) of the function to be minimized for selecting the optimal resolution.}

\item{showOpt}{(logical) plot the result.}
}
\value{
It returns a list with: i) a matrix reporting the values of the function to be minimized, along with the corresponding resolution; ii) the optimal resolution.
}
\description{
This function finds the optimal resolution to perform the uniform environmental sampling.
It requires an sf object having point geometry given by the PC-scores values (the first two axes of a PCA computed on the environmental variables of interest),
and a set of proposed resolution for the sampling grid to test.
For a given resolution, the function iterates across each cell and: i) finds the centroid of the PC-scores within the cell;
ii) computes the average (squared) euclidean distance between each PC-score and their centroid.
Finally, the cell-specific distances are averaged by the number of cells.
It returns a list with: i) a matrix reporting the values of the function to be minimized, along with the corresponding resolution; ii) the optimal resolution.
The function also provides a plot displaying the function values for each resolution, allowing the user to select a resolution on their own.
}
