% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimRes.R
\name{optimRes}
\alias{optimRes}
\title{Get optimal resolution of the sampling grid}
\usage{
optimRes(sdf, grid.res, perc.thr = 10, cr = 1, showOpt = TRUE)
}
\arguments{
\item{sdf}{an sf object having point geometry given by the PC-scores values}

\item{grid.res}{(integer) a vector of resolutions to be tested, i.e seq(1,100, by=1)}

\item{perc.thr}{rate of change (expressed in percentage) of the function to be minimized for selecting the optimal resolution.}

\item{cr}{(integer) number of cores for parallel computing.}

\item{showOpt}{(logical) plot the result.}
}
\value{
It returns a list with: i) a matrix reporting the values of the function to be minimized, along with the corresponding resolution; ii) the optimal resolution.
}
\description{
\code{optimRes} identifies the optimal resolution of the sampling grid to be used to perform the uniform environmental sampling.
The function requires an sf object of geometry type “POINT”. Points are the scores of the first two axes of a principal component analysis performed on the correlation matrix of a set of environmental variables. A set of candidate resolutions to be test to find the optimal resolution of the sampling grid must be provided. For each candidate resolution, \code{optimRes} computes a measure summarizing the average (squared) Euclidean distance between points within each cell and the centroid of the convex hull built around the points. This measure is then compared across sampling grids with increasing resolutions (i.e., number of cells). The best resolution is selected as that allowing to find the best trade-off between resolution and average distance among points within cells (i.e., resolution that allows uniformly sampling the environmental space without overfitting it). More specifically, the best grid is the one whose resolution is just below that which would not allow the average distance among pseudo-absences to be reduced by more than 10\% (10\% is the default value, but another threshold can be set by the user).
\code{optimRes} returns a list of length 2 with a matrix reporting the measure computed for each sampling grid at the corresponding resolution, and the selected optimal resolution.
The function also provides a plot displaying the function values for each resolution, allowing the user to select a resolution on their own.
}
