% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_predict.R
\name{pca_predict}
\alias{pca_predict}
\title{Predict pca}
\usage{
pca_predict(data, model, nPC)
}
\arguments{
\item{data}{A RasterStack, RasterBrick or a SpatRaster object comprising the variables describing the environmental space.}

\item{model}{\code{princomp} object.}

\item{nPC}{Integer. Number of PCA components to return.}
}
\description{
Predict pca
}
