% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uesampling.R
\name{uesampling}
\alias{uesampling}
\title{Uniform sampling of the environmental space}
\usage{
uesampling(
  sdf,
  grid.res,
  n.tr = 5,
  sub.ts = FALSE,
  n.ts = FALSE,
  plot_proc = TRUE
)
}
\arguments{
\item{sdf}{an sf object having point geometry given by the PC-scores values}

\item{grid.res}{(integer) resolution of the sampling grid. The resolution can be arbitrarily selected or defined using the Optim_res() function.}

\item{n.tr}{(integer) number of background points for the training dataset to sample in each cell of the sampling grid}

\item{sub.ts}{(logical) sample the validation background points}

\item{n.ts}{(integer; optional) number of background points for the testing dataset to sample in each cell of the sampling grid. sub.ts argument must be TRUE.}

\item{plot_proc}{(logical) plot progress of the sampling}
}
\value{
A spatial point data frame with the coordinates of the sampled background points both in the geographical and environmental space
}
\description{
This function allows to perform a uniform sampling of the environmental space.
It requires an sf object having point geometry given by the PC-scores values (the first two axes of a PCA computed on the environmental variables of interest), and the
bounding box of the environmental space (defined by the extent of the PC-scores values).
A resolution of the sampling grid must be provided by the user. The resolution can be arbitrarily selected or
or defined using the Optim_res() function.
The functions allows to retrieve both a training and a testing dataset (optional) of uniformely sampled background points.
The user must provide suitable numbers of background points to be sampled in each cell of the sampling grid
(n.tr: background points for the training dataset; n.ts: background points for the testing dataset).
Notice that the final number of sampled background points depends on the PC-scores configuration in the environmental space.
}
