% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Worldclim_tmp.R
\docType{data}
\name{Worldclim_tmp}
\alias{Worldclim_tmp}
\title{A subset of WorldClim bioclimatic variables}
\format{
A data frame obtained from a SpatRaster with 1080 rows, 2160 columns, and 6 layers, namely:"bio1"  "bio3"  "bio9"  "bio12" "bio13" "bio15"
}
\source{
\code{geodata::worldclim_global(var='bio', res=10, path=getwd())[[c(1, 3,9, 12, 13, 15)]]}
}
\usage{
data(Worldclim_tmp)
}
\description{
A subset of WorldClim bioclimatic variables cropped on the Central and Western Europe.
}
\keyword{datasets}
