% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimRes.R
\name{optimRes}
\alias{optimRes}
\title{Get optimal resolution of the sampling grid}
\usage{
optimRes(sdf, grid.res, perc.thr = 10, cr = 1, showOpt = TRUE)
}
\arguments{
\item{sdf}{an sf object having point geometry given by the PC-scores values}

\item{grid.res}{(integer) a vector of resolutions to be tested, i.e seq(1,100, by=1)}

\item{perc.thr}{rate of change (expressed in percentage) of the function to be minimized for selecting the optimal resolution.}

\item{cr}{(integer) number of cores for parallel computing. The default cluster type is PSOCK.}

\item{showOpt}{(logical) plot the result.}
}
\value{
It returns a list with: i) a matrix reporting the values of the function to be minimized, along with the corresponding resolution; ii) the optimal resolution.
}
\description{
\code{optimRes} identifies the optimal resolution of the sampling grid to be used to perform the uniform environmental sampling.
The function requires an sf object of geometry type “POINT”. Points here are the scores of the first two axes of a principal component analysis performed on the correlation matrix of a set of environmental variables. A set of candidate resolutions must be provided to find the optimal resolution of the sampling grid. For each candidate resolution, \code{optimRes} computes a metric summarizing the average squared Euclidean distance between points within each cell and the centroid of the convex hull built around the points. This metric is then compared across different sampling grids with increasing resolution (i.e., number of cells). The best resolution is selected as the one allowing finding the best trade-off between the number of cells and the average distance among points within cells. In other words, this is the resolution allowing for uniformly sampling the environmental space without overfitting it. This is obtained by choosing when the average distance among pseudo-absences cannot be reduced by more than 10\% (10\% is the default value, but another threshold can be set by the user).
\code{optimRes} returns a list of length 2 with a matrix reporting the measure computed for each sampling grid at the corresponding resolution and the selected optimal resolution.
The function also provides a plot displaying the function values for each resolution, allowing the user to select a resolution on their own.
}
\details{
In case the function returns NA as the optimal resolution: i) increase the range of \code{grid.res}, ii) increase \code{perc.thr}.
}
