% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh.inspect.R
\name{thresh.inspect}
\alias{thresh.inspect}
\title{Inspect the effect of the kernel threshold parameter on the environmental space partitioning}
\usage{
thresh.inspect(env.rast, pres = NULL, thres = 0.75, H = NULL)
}
\arguments{
\item{env.rast}{A RasterStack, RasterBrick or a SpatRaster object comprising the variables describing the environmental space.}

\item{pres}{A SpatialPointsDataframe, a SpatVector or an sf object including the presence-only observations of the species of interest.}

\item{thres}{(double) This value or vector of values identifies the quantile value used to specify the boundary of the kernel density estimate (default \code{thres=0.75} ). Thus, probability values higher than the threshold should indicate portions of the multivariate space likely associated with presence points.}

\item{H}{The kernel bandwidth (i.e., the width of the kernel density function that defines its shape) excluding the portion of the environmental space associated with environmental conditions likely suitable for the species. It can be either defined by the user or automatically estimated by \code{paSampling} via \code{ks::Hpi}.}
}
\value{
A ggplot2 object showing how the environmental space is partitioned accordingly to the selected \code{thres} values.
}
\description{
\code{thresh.inspect} allows inspecting the influence of the kernel density threshold on the partition of the environmental space performed by the \code{paSampling} function.
The environmental space is partitioned as follows: a kernel density estimate of the species presence observations identifies and excludes the portion of the environmental space associated with environmental conditions likely suitable for the species to establish using a threshold specified by the user, which can be visually inspected by the user using \code{thresh.inspect}.
The bandwidth of the kernel can be estimated automatically or defined by the user, allowing to account for meta-population dynamics (e.g. sink-source populations).
}
