% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniformSampling.R
\name{uniformSampling}
\alias{uniformSampling}
\title{Uniform sampling of the environmental space}
\usage{
uniformSampling(
  sdf,
  grid.res,
  n.tr = 5,
  n.prev = NULL,
  sub.ts = FALSE,
  n.ts = 5,
  plot_proc = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sdf}{an sf object having point geometry given by the PC-scores values}

\item{grid.res}{(integer) resolution of the sampling grid. The resolution can be arbitrarily selected or defined using the \code{optimRes()} function.}

\item{n.tr}{(integer; optional) number of expected points given a certain prevalence threshold for the training dataset.}

\item{n.prev}{(double) sample prevalence}

\item{sub.ts}{(logical) sample the validation points}

\item{n.ts}{(integer; optional) number of  points for the testing dataset to sample in each cell of the sampling grid. sub.ts argument must be TRUE.}

\item{plot_proc}{(logical) plot progress of the sampling}

\item{verbose}{(logical) Print verbose}
}
\value{
A spatial point data frame with the coordinates of the sampled points both in the geographical and environmental space
}
\description{
\code{uniformSampling} performs the uniform sampling of the environmental space.
It requires an sf object of geometry type “POINT”. Points are the scores of the first two axes of a principal component analysis performed on the correlation matrix of a set of environmental variables. The resolution of the sampling grid used to systematically collect observations within the environmental space must be provided. The resolution can be either set to an arbitrary value by the user or selected using the \code{optimRes()} function.
\code{uniformSampling} allows collecting observations to populate both training and (optional) a testing dataset.
In both cases, the user must provide a number of points to be sampled in each cell of the sampling grid (\code{n.tr:} points for the training dataset; \code{n.ts}: points for the testing dataset).
Note that the number of pseudo-absences eventually sampled depends on the spatial configuration of the points within the environmental space. Usually, some cells of the sampling grid are empty (i.e., do not include points), as points are not covering the whole extent of the environmental space. For this reason, the number of pseudo-absences eventually collected is likely to be lower than the product between the resolution of the sampling grid and \code{n.tr}(or \code{n.ts}).
}
