% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Worldclim_tmp.R
\docType{data}
\name{Worldclim_tmp}
\alias{Worldclim_tmp}
\title{A subset of WorldClim bioclimatic variables}
\format{
A data frame obtained from a SpatRaster with 1080 rows, 2160 columns, and 5 layers, namely: wc2.1_10m_bio_4, wc2.1_10m_bio_3, wc2.1_10m_bio_14, wc2.1_10m_bio_9, wc2.1_10m_bio_15.
}
\source{
\code{geodata::worldclim_global(var='bio', res=10, path=getwd())[[c(4,  3, 14,  9, 15)]]}
}
\usage{
data(Worldclim_tmp)
}
\description{
A subset of WorldClim bioclimatic variables cropped on the Central and Western Europe.
}
\keyword{datasets}
